Guest can share news through social networks
Added template with share buttons.  This template can be rendered
from any controller or template.  Url, title and description can
be sent to this template from any controller.  Added new web
steps for searching in DOM because I need to do it for testing if
div with share buttons exists.

Story: 146
Ticket: 178
Change-Id: I9fedc45d79b05db3d31785170d649d78f2ed7c93
Reviewed-on: http://gerrit.altoros.corp/63
Tested-by: Eugene Melnikov <melnikaite.melnikaite@gmail.com>
Reviewed-by: Sergey Sergyenko <sergey.sergyenko@gmail.com>